backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
openai-whisper==20231117
groq>=0.4.0
requests==2.31.0
aiofiles==23.2.1
certifi
